refactor: consolidate homebrew release workflow generation into a single function
