+++ crates/forge_ci/src/workflows/release_homebrew.rs
@@ -1,10 +1,29 @@
 use gh_workflow_tailcall::generate::Generate;
+use gh_workflow_tailcall::*;
 
-use crate::jobs;
+use crate::jobs::{self, create_build_release_job_for_publishing, create_homebrew_release_job};
+use crate::matrix;
 
 /// Generate homebrew release workflow
 pub fn generate_homebrew_workflow() {
-    let homebrew_workflow = jobs::create_homebrew_workflow();
+    let mut homebrew_workflow = Workflow::default()
+        .name("Homebrew Release")
+        .on(Event {
+            release: Some(Release { types: vec![ReleaseType::Published] }),
+            ..Event::default()
+        })
+        .permissions(
+            Permissions::default()
+                .contents(Level::Write)
+                .pull_requests(Level::Write),
+        );
+
+    let build_job = create_build_release_job_for_publishing(matrix::create_matrix());
+    let homebrew_release_job = create_homebrew_release_job().add_needs(build_job.clone());
+
+    homebrew_workflow = homebrew_workflow
+        .add_job("build-release", build_job)
+        .add_job("homebrew_release", homebrew_release_job);
 
     Generate::new(homebrew_workflow)
         .name("release-homebrew.yml")
